Bullet Projectile
I added projectile movement in the Bullet, which gives it an initial
velocity. However, it seems that there is an additional force acting
on the bullet causing it to be pushed in different directions from
the set direction. The PlayerCapsule Event graph sets the direction
and location of the projectile, while the Bullet projectile movement
is modified to change how gravity and velocity affects it.